add header for toggle-theme-icon
