+++ src/components/icons/HomeIcons.tsx
@@ -2,7 +2,6 @@ import blogConfig from 'blog.config'
 import { useContext } from 'react'
 import Link from 'next/link'
 import css from 'styled-jsx/css'
-// import { ThemeContext } from '../../pages/_app'
 import { ThemeContext } from '../../hooks/theme'
 import { IconContext } from 'react-icons'
 import { WantedlySvg } from './index'
